Fix vllm Base and Chat Completion template
