+++ python/huggingfaceserver/huggingfaceserver/vllm_completions.py
@@ -14,19 +14,36 @@
 
 import time
 import torch
+import asyncio
+from http import HTTPStatus
 from openai.types import Completion, CompletionChoice, CompletionUsage
+from openai.types.chat import ChatCompletionMessageParam
 from vllm.sampling_params import SamplingParams
 from vllm.utils import random_uuid
-from typing import AsyncGenerator, AsyncIterator, List, Tuple
-from fastapi import Request
+from typing import (
+    AsyncGenerator,
+    AsyncIterator,
+    Dict,
+    Iterable,
+    List,
+    Optional,
+    Tuple,
+    Union,
+)
 from vllm.outputs import RequestOutput
 from vllm.entrypoints.openai.serving_completion import (
     parse_prompt_format,
     merge_async_iterators,
 )
-from .completions_utils import OpenAIServing
 from vllm.engine.async_llm_engine import AsyncLLMEngine
-from kserve.protocol.rest.openai.types.openapi import CreateCompletionRequest
+from vllm.transformers_utils.tokenizer import get_tokenizer
+from vllm.sequence import Logprob
+from kserve.protocol.rest.openai.types.openapi import (
+    CreateCompletionRequest,
+    ErrorResponse,
+    Error,
+    Logprobs,
+)
 
 
 def to_sampling_params(request: CreateCompletionRequest):
@@ -62,14 +79,30 @@ def to_sampling_params(request: CreateCompletionRequest):
     )
 
 
-class OpenAIServingCompletion(OpenAIServing):
+class OpenAIServingCompletion:
 
     def __init__(self, engine: AsyncLLMEngine, served_model: str):
-        super().__init__(engine=engine, served_model=served_model)
+        self.engine = engine
+        self.served_model = served_model
 
-    async def create_completion(
-        self, request: CreateCompletionRequest, raw_request: Request
-    ):
+        self.max_model_len = 0
+        self.tokenizer = None
+
+        try:
+            event_loop = asyncio.get_running_loop()
+        except RuntimeError:
+            event_loop = None
+
+        if (
+            event_loop is not None and event_loop.is_running()
+        ):  # If the current is instanced by Ray Serve, there is already a running event loop
+            event_loop.create_task(self._post_init())
+        else:  # When using single vLLM without engine_use_ray
+            loop = asyncio.new_event_loop()
+            loop.run_until_complete(self._post_init())
+            loop.close()
+
+    async def create_completion(self, request: CreateCompletionRequest):
         """Completion API similar to OpenAI's API.
 
         See https://platform.openai.com/docs/api-reference/completions/create
@@ -132,7 +165,6 @@ class OpenAIServingCompletion(OpenAIServing):
         if stream:
             return self.completion_stream_generator(
                 request,
-                raw_request,
                 result_generator,
                 request_id,
                 created_time,
@@ -144,10 +176,6 @@ class OpenAIServingCompletion(OpenAIServing):
         final_res_batch: RequestOutput = [None] * len(prompts)
         try:
             async for i, res in result_generator:
-                if await raw_request.is_disconnected():
-                    # Abort the request if the client disconnects.
-                    await self.engine.abort(f"{request_id}-{i}")
-                    return self.create_error_response("Client disconnected")
                 final_res_batch[i] = res
             response = self.request_output_to_completion_response(
                 final_res_batch, request, request_id, created_time, model_name
@@ -160,7 +188,6 @@ class OpenAIServingCompletion(OpenAIServing):
     async def completion_stream_generator(
         self,
         request: CreateCompletionRequest,
-        raw_request: Request,
         result_generator: AsyncIterator[Tuple[int, RequestOutput]],
         request_id: str,
         created_time: int,
@@ -173,11 +200,6 @@ class OpenAIServingCompletion(OpenAIServing):
 
         try:
             async for prompt_idx, res in result_generator:
-
-                # Abort the request if the client disconnects.
-                if await raw_request.is_disconnected():
-                    await self.engine.abort(f"{request_id}-{prompt_idx}")
-                    raise StopAsyncIteration()
 
                 for output in res.outputs:
                     i = output.index + prompt_idx * request.n
@@ -317,3 +339,114 @@ class OpenAIServingCompletion(OpenAIServing):
             choices=choices,
             usage=usage,
         )
+
+    def apply_chat_template(self, messages: Iterable[ChatCompletionMessageParam]):
+        return self.tokenizer.apply_chat_template(conversation=messages, tokenize=False)
+
+    def create_error_response(
+        self,
+        message: str,
+        err_type: str = "BadRequestError",
+        status_code: HTTPStatus = HTTPStatus.BAD_REQUEST,
+    ) -> ErrorResponse:
+
+        error = Error(
+            message=message, type=err_type, param="", code=str(status_code.value)
+        )
+        return ErrorResponse(error=error)
+
+    def create_streaming_error_response(
+        self,
+        message: str,
+        err_type: str = "BadRequestError",
+        status_code: HTTPStatus = HTTPStatus.BAD_REQUEST,
+    ) -> ErrorResponse:
+        return self.create_error_response(
+            message=message, err_type=err_type, status_code=status_code
+        )
+
+    async def _post_init(self):
+        engine_model_config = await self.engine.get_model_config()
+        self.max_model_len = engine_model_config.max_model_len
+
+        # A separate tokenizer to map token IDs to strings.
+        self.tokenizer = get_tokenizer(
+            engine_model_config.tokenizer,
+            tokenizer_mode=engine_model_config.tokenizer_mode,
+            trust_remote_code=engine_model_config.trust_remote_code,
+        )
+
+    async def _check_model(self, request) -> Optional[ErrorResponse]:
+        if request.model == self.served_model:
+            return
+        return self.create_error_response(
+            message=f"The model `{request.model}` does not exist.",
+            err_type="NotFoundError",
+            status_code=HTTPStatus.NOT_FOUND,
+        )
+
+    def _validate_prompt_and_tokenize(
+        self,
+        request: Union[CreateCompletionRequest],
+        prompt: Optional[str] = None,
+        prompt_ids: Optional[List[int]] = None,
+    ) -> List[int]:
+        if not (prompt or prompt_ids):
+            raise ValueError("Either prompt or prompt_ids should be provided.")
+        if prompt and prompt_ids:
+            raise ValueError("Only one of prompt or prompt_ids should be provided.")
+
+        input_ids = (
+            prompt_ids if prompt_ids is not None else self.tokenizer(prompt).input_ids
+        )
+        token_num = len(input_ids)
+
+        if request.max_tokens is None:
+            request.max_tokens = self.max_model_len - token_num
+
+        if token_num + request.max_tokens > self.max_model_len:
+            raise ValueError(
+                f"This model's maximum context length is "
+                f"{self.max_model_len} tokens. However, you requested "
+                f"{request.max_tokens + token_num} tokens "
+                f"({token_num} in the messages, "
+                f"{request.max_tokens} in the completion). "
+                f"Please reduce the length of the messages or completion.",
+            )
+        else:
+            return input_ids
+
+    def _create_logprobs(
+        self,
+        token_ids: List[int],
+        top_logprobs: Optional[List[Optional[Dict[int, Logprob]]]] = None,
+        num_output_top_logprobs: Optional[int] = None,
+        initial_text_offset: int = 0,
+    ) -> Logprobs:
+        """Create OpenAI-style logprobs."""
+        logprobs = Logprobs()
+        last_token_len = 0
+        if num_output_top_logprobs:
+            logprobs.top_logprobs = []
+        for i, token_id in enumerate(token_ids):
+            step_top_logprobs = top_logprobs[i]
+            if step_top_logprobs is not None:
+                token_logprob = step_top_logprobs[token_id].logprob
+            else:
+                token_logprob = None
+            token = step_top_logprobs[token_id].decoded_token
+            logprobs.tokens.append(token)
+            logprobs.token_logprobs.append(token_logprob)
+            if len(logprobs.text_offset) == 0:
+                logprobs.text_offset.append(initial_text_offset)
+            else:
+                logprobs.text_offset.append(logprobs.text_offset[-1] + last_token_len)
+            last_token_len = len(token)
+
+            if num_output_top_logprobs:
+                logprobs.top_logprobs.append(
+                    {p.decoded_token: p.logprob for i, p in step_top_logprobs.items()}
+                    if step_top_logprobs
+                    else None
+                )
+        return logprobs
